Readme fixes, move plugin to its own file
