+++ tests/contrib/openai_agents/test_openai_replay.py
@@ -1,10 +1,4 @@
-from contextlib import (
-    AbstractAsyncContextManager,
-    AbstractContextManager,
-    asynccontextmanager,
-)
 from pathlib import Path
-from typing import AsyncGenerator
 
 import pytest
